requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest>=8.0.0
pytest-qt>=4.4.0
pytest-cov>=4.1.0

# Code quality
ruff>=0.1.0
mypy>=1.8.0

# Type stubs
PyQt6-stubs>=6.6.0
